Created the Login Page
I created the login page. I decided I'm not going to have a sign up page. I'm going to hard code the accounts as objects. Now I need to create the banking page. Then I need to create the javascript logic to connect the login page with the banking page.